Refactor Cluster Agents do it matches the docs
