+++ cluster_agents.go
@@ -22,18 +22,18 @@ import (
 	"time"
 )
 
-// ClusterAgentsService handles communication with the `cluster_agents` related methods
-// of the GitLab API.
+// ClusterAgentsService handles communication with the cluster agents related
+// methods of the GitLab API.
 //
 // GitLab API docs: https://docs.gitlab.com/ee/api/cluster_agents.html
 type ClusterAgentsService struct {
 	client *Client
 }
 
-// ClusterAgent represents a GitLab Agent for Kubernetes
+// Agent represents a GitLab agent for Kubernetes.
 //
 // GitLab API docs: https://docs.gitlab.com/ee/api/cluster_agents.html
-type ClusterAgent struct {
+type Agent struct {
 	ID              int           `json:"id"`
 	Name            string        `json:"name"`
 	CreatedAt       *time.Time    `json:"created_at"`
@@ -51,19 +51,41 @@ type ConfigProject struct {
 	CreatedAt         *time.Time `json:"created_at"`
 }
 
-func (a ClusterAgent) String() string {
+func (a Agent) String() string {
 	return Stringify(a)
 }
 
-// ListClusterAgentsOptions represents the available ClusterAgents() options.
+// AgentToken represents a GitLab agent token.
 //
-// GitLab API docs: https://docs.gitlab.com/ee/api/cluster_agents.html#list-the-agents-for-a-project
-type ListClusterAgentsOptions ListOptions
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#list-tokens-for-an-agent
+type AgentToken struct {
+	ID              int        `json:"id"`
+	Name            string     `json:"name"`
+	Description     string     `json:"description"`
+	AgentID         int        `json:"agent_id"`
+	Status          string     `json:"status"`
+	CreatedAt       *time.Time `json:"created_at"`
+	CreatedByUserID int        `json:"created_by_user_id"`
+	LastUsedAt      *time.Time `json:"last_used_at"`
+	Token           string     `json:"token"`
+}
 
-// ListClusterAgents returns a list of Agents for Kubernetes registered for the project
+func (a AgentToken) String() string {
+	return Stringify(a)
+}
+
+// ListProjectAgentsOptions represents the available ListProjectAgents() options.
 //
-// GitLab API docs: https://docs.gitlab.com/ee/api/cluster_agents.html#list-the-agents-for-a-project
-func (s *ClusterAgentsService) ListClusterAgents(pid interface{}, opt *ListClusterAgentsOptions, options ...RequestOptionFunc) ([]*ClusterAgent, *Response, error) {
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#list-the-agents-for-a-project
+type ListProjectAgentsOptions ListOptions
+
+// ListProjectAgents returns a list of agents registered for the project.
+//
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#list-the-agents-for-a-project
+func (s *ClusterAgentsService) ListProjectAgents(pid interface{}, opt *ListProjectAgentsOptions, options ...RequestOptionFunc) ([]*Agent, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -75,19 +97,20 @@ func (s *ClusterAgentsService) ListClusterAgents(pid interface{}, opt *ListClust
 		return nil, nil, err
 	}
 
-	var t []*ClusterAgent
-	resp, err := s.client.Do(req, &t)
+	var as []*Agent
+	resp, err := s.client.Do(req, &as)
 	if err != nil {
 		return nil, resp, err
 	}
 
-	return t, resp, err
+	return as, resp, err
 }
 
-// GetClusterAgent gets an Agent for Kubernetes by ID
+// GetAgent gets a single agent details.
 //
-// GitLab API docs: https://docs.gitlab.com/ee/api/cluster_agents.html#get-details-about-an-agent
-func (s *ClusterAgentsService) GetClusterAgent(pid interface{}, id int, options ...RequestOptionFunc) (*ClusterAgent, *Response, error) {
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#get-details-about-an-agent
+func (s *ClusterAgentsService) GetAgent(pid interface{}, id int, options ...RequestOptionFunc) (*Agent, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -99,28 +122,29 @@ func (s *ClusterAgentsService) GetClusterAgent(pid interface{}, id int, options 
 		return nil, nil, err
 	}
 
-	t := new(ClusterAgent)
-	resp, err := s.client.Do(req, t)
+	a := new(Agent)
+	resp, err := s.client.Do(req, a)
 	if err != nil {
 		return nil, resp, err
 	}
 
-	return t, resp, err
+	return a, resp, err
 }
 
-// RegisterClusterAgentOptions represents the available RegisterClusterAgent() options.
+// RegisterProjectAgentOptions represents the available RegisterProjectAgent()
+// options.
 //
 // GitLab API docs:
 // https://docs.gitlab.com/ee/api/cluster_agents.html#register-an-agent-with-a-project
-type RegisterClusterAgentOptions struct {
+type RegisterProjectAgentOptions struct {
 	Name *string `url:"name,omitempty" json:"name,omitempty"`
 }
 
-// RegisterClusterAgent registers a new Agent for Kubernetes for a project
+// RegisterProjectAgent registers an agent to the project.
 //
 // GitLab API docs:
 // https://docs.gitlab.com/ee/api/cluster_agents.html#register-an-agent-with-a-project
-func (s *ClusterAgentsService) RegisterClusterAgent(pid interface{}, opt *RegisterClusterAgentOptions, options ...RequestOptionFunc) (*ClusterAgent, *Response, error) {
+func (s *ClusterAgentsService) RegisterProjectAgent(pid interface{}, opt *RegisterProjectAgentOptions, options ...RequestOptionFunc) (*Agent, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -132,25 +156,134 @@ func (s *ClusterAgentsService) RegisterClusterAgent(pid interface{}, opt *Regist
 		return nil, nil, err
 	}
 
-	t := new(ClusterAgent)
-	resp, err := s.client.Do(req, t)
+	a := new(Agent)
+	resp, err := s.client.Do(req, a)
 	if err != nil {
 		return nil, resp, err
 	}
 
-	return t, resp, err
+	return a, resp, err
 }
 
-// DeleteClusterAgent deletes a registered Agent for Kubernetes
+// DeleteAgent deletes an existing agent registration.
 //
 // GitLab API docs:
 // https://docs.gitlab.com/ee/api/cluster_agents.html#delete-a-registered-agent
-func (s *ClusterAgentsService) DeleteClusterAgent(pid interface{}, id int, options ...RequestOptionFunc) (*Response, error) {
+func (s *ClusterAgentsService) DeleteAgent(pid interface{}, id int, options ...RequestOptionFunc) (*Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, err
 	}
 	uri := fmt.Sprintf("projects/%s/cluster_agents/%d", PathEscape(project), id)
+
+	req, err := s.client.NewRequest(http.MethodDelete, uri, nil, options)
+	if err != nil {
+		return nil, err
+	}
+
+	return s.client.Do(req, nil)
+}
+
+// ListAgentTokensOptions represents the available ListAgentTokens() options.
+//
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#list-tokens-for-an-agent
+type ListAgentTokensOptions ListOptions
+
+// ListAgentTokens returns a list of tokens for an agent.
+//
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#list-tokens-for-an-agent
+func (s *ClusterAgentsService) ListAgentTokens(pid interface{}, aid int, opt *ListAgentTokensOptions, options ...RequestOptionFunc) ([]*AgentToken, *Response, error) {
+	project, err := parseID(pid)
+	if err != nil {
+		return nil, nil, err
+	}
+	uri := fmt.Sprintf("projects/%s/cluster_agents/%d/tokens", PathEscape(project), aid)
+
+	req, err := s.client.NewRequest(http.MethodGet, uri, opt, options)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	var ats []*AgentToken
+	resp, err := s.client.Do(req, &ats)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return ats, resp, err
+}
+
+// GetAgentToken gets a single agent token.
+//
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#get-a-single-agent-token
+func (s *ClusterAgentsService) GetAgentToken(pid interface{}, aid int, id int, options ...RequestOptionFunc) (*AgentToken, *Response, error) {
+	project, err := parseID(pid)
+	if err != nil {
+		return nil, nil, err
+	}
+	uri := fmt.Sprintf("projects/%s/cluster_agents/%d/tokens/%d", PathEscape(project), aid, id)
+
+	req, err := s.client.NewRequest(http.MethodGet, uri, nil, options)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	at := new(AgentToken)
+	resp, err := s.client.Do(req, at)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return at, resp, err
+}
+
+// CreateAgentTokenOptions represents the available CreateAgentToken() options.
+//
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#create-an-agent-token
+type CreateAgentTokenOptions struct {
+	Name        *string `url:"name,omitempty" json:"name,omitempty"`
+	Description *string `url:"description,omitempty" json:"description,omitempty"`
+}
+
+// CreateAgentToken creates a new token for an agent.
+//
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#create-an-agent-token
+func (s *ClusterAgentsService) CreateAgentToken(pid interface{}, aid int, opt *CreateAgentTokenOptions, options ...RequestOptionFunc) (*AgentToken, *Response, error) {
+	project, err := parseID(pid)
+	if err != nil {
+		return nil, nil, err
+	}
+	uri := fmt.Sprintf("projects/%s/cluster_agents/%d/tokens", PathEscape(project), aid)
+
+	req, err := s.client.NewRequest(http.MethodPost, uri, opt, options)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	at := new(AgentToken)
+	resp, err := s.client.Do(req, at)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return at, resp, err
+}
+
+// RevokeAgentToken revokes an agent token.
+//
+// GitLab API docs:
+// https://docs.gitlab.com/ee/api/cluster_agents.html#revoke-an-agent-token
+func (s *ClusterAgentsService) RevokeAgentToken(pid interface{}, aid int, id int, options ...RequestOptionFunc) (*Response, error) {
+	project, err := parseID(pid)
+	if err != nil {
+		return nil, err
+	}
+	uri := fmt.Sprintf("projects/%s/cluster_agents/%d/tokens/%d", PathEscape(project), aid, id)
 
 	req, err := s.client.NewRequest(http.MethodDelete, uri, nil, options)
 	if err != nil {
